+++ src/index.ts
@@ -17,7 +17,11 @@ export function getMetadata<TOptions extends MediaMetadataPublicFields>(
   >;
 }
 
-/** Returns the artwork of specified media file from its uri. */
+/**
+ * Returns the artwork of the specified media file from its uri. Has ~5MB
+ * cap on what's returned (image must be at most 3.75MB due to 33-37%
+ * increase to base64).
+ */
 export function getArtwork(uri: string): Promise<string | null> {
   return MetadataRetriever.getArtwork(uri);
 }